+++ resources/views/admin/hotspot-asset/create.php
@@ -3,18 +3,18 @@
 declare(strict_types=1);
 
 /**
- * @see HotspotAssetController::actionIndex()
+ * @see AbstractAssetController::actionCreate()
  *
  * @var View $this
  * @var Hotspot $model
- * @var AssetArrayDataProvider $provider
+ * @var FileActiveDataProvider $provider
  */
 
 use Hirtz\Cms\Hotspot\Models\Hotspot;
-use Hirtz\Cms\Hotspot\Modules\Admin\Controllers\HotspotAssetController;
-use Hirtz\Cms\Hotspot\Modules\Admin\Widgets\Grids\HotspotAssetGridView;
 use Hirtz\Cms\Hotspot\Modules\Admin\Widgets\Navs\HotspotSubmenu;
-use Hirtz\Media\Modules\Admin\Data\AssetArrayDataProvider;
+use Hirtz\Media\Modules\Admin\Controllers\AbstractAssetController;
+use Hirtz\Media\Modules\Admin\Data\FileActiveDataProvider;
+use Hirtz\Media\Modules\Admin\Widgets\Grids\FileGridView;
 use Hirtz\Skeleton\I18n\Lang;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Grids\GridContainer;
@@ -24,6 +24,9 @@ $this->title(Lang::t('media', 'ASSET_MODEL_LABEL'));
 echo HotspotSubmenu::make()
     ->hotspot($model);
 
+$this->addBreadcrumb(Lang::t('media', 'COMMON_LINK_ASSETS'));
+
 echo GridContainer::make()
-    ->grid(HotspotAssetGridView::make()
+    ->grid(FileGridView::make()
+        ->model($model)
         ->provider($provider));